OpenGL = garbage
Here's why.

So for some reason some cunt decided to make Y=0 in UV coordinates be the BOTTOM of the texture.
Whoever came up with this stupid decision, I curse you and your programming carreer over the hours I've spent debugging stupid issues related to this.
The problem I came across was that I was sending in images where Y=0 is at the top. I was wondering why my images looked correct and why framebuffers were flipped around, and here's why:
When you send an image into a texture, the first scanline of the image (that is, the one that begins at byte 0) will be assigned Y=0 in UV coordinates, going upwards as it scans more lines.
Now, in NDC, the image will appear flipped, but then the orthographic projection flips the coordinate space around, which means that the coordinates of the image will indeed be correct.
But what happens if you try to draw a framebuffer texture instead? Well, shit goes downhill, and now you have to deal with the fact that for OpenGL Y=0 is at the bottom of the texture. You flip the UVs and call it a day. So framebuffeer textures will be flipped around, but your good ol' normal image textures won't.
Because of this, when sending in external data to fill the framebuffer with (using glTexSubImage), you need to flip all the pixels 'round on the Y axis, as well as the Y position you're plopping the texture sub image on.

Jesus fuck, this took me hours to debug. So here's to future generations HOPEFULLY not having to debug this. At least I know I won't have to.

Take care.