rather simple clipping
I think with this change the OpenGL backend is mostly done.
The clipping implementation in this commit is very primitive and does not handle transformations at all. NetCanv doesn't use clipping that heavily though, and only through paws, which does not translate the transform matrix before using clip(), so it hopefully isn't going to be a problem. Hopefully.