added a magic multiplier that makes #FFFFFF possible
This multiplier was found purely by experimentation and println!.
Multiplying the value by it before converting to Oklab (or dividing when converting from Oklab)
causes #FFFFFF possible.

I found this multiplier by raising the maximum possible V component in the GUI, to find out that
there actually are correct colors beyond my assumed limit of 1.0. Then I simply dragged the
handle on the color canvas to the minimum Y position that produced #FFFFFF.