rewrite socket system to use tokio
Third time's the charm.
This version uses tokio + async because of `select!`. It isn't possible to select between two
tasks using _only_ the standard library, which is a shame but it is what it is.
Besides I'm not a big fan of nysa anymore. I think it has some fundamental flaws that are better solved using an actor model like the one used in this implementation.