Remove a peer when relay tells us there is no such peer
This fixes an error where if somehow relay's state and host's state are
desynchronized and host thinks there is a peer that already left,
leading to host sending a message to them and relay responding that
the peer doesn't exists, which causes the host to leave the room.

See a comment in src/net/peer.rs to see why this is fine.