Do the anti-aliasing hack before comparing points a and b
This fixes a bug where a brush with an odd number of pixels could draw
one more pixel than it should.

When the user is drawing slowly, the code considers points a and b to
be equal and draws a point there. But when the user is drawing faster,
the points a and b aren't considered equal, and draws a line instead.

The problem was, that we do a hack to solve anti-aliasing when drawing 1
pixel lines. But, the hack *only* applies when points a and b aren't
considered equal. So, in the first case, the hack doesn't apply and
points a and b are different than in the second case.

The fix is to do the hack first, solving the issue.