Select existing color from the palette when picking the same color (#236)
When we draw something using color from the palette, change the selected
color, and pick the color we drew with the eyedropper, it overrides
currently selected color in the palette with the picked color. This
commit changes this behavior to, if the palette has the picked color
already, not override the currently selected color, but change it to the
color in the palette instead.