Send deselect packet when deselecting after paste
When user is doing actions in this order: paste image, move, paste
image, move, deselect; we are sending those packets: Paste, Rect, Paste,
Rect, Deselect. However, this causes issues, because selection
rectangle can get updated before we deselect it, thus placing it at
wrong position.

The previous fix was to update rectangle before sending Paste packet,
and then "pair" the rectangle with Paste packet on the peers side
for later usage. But this resulted in more issues, like deselecting with
wrong rectangle or at wrong position.

In this fix, we think instead in what client does, and not what user
does. For the client, the actions are: paste image, move, _deselect_
(because we have a selection), paste image, move and deselect. So, when
deselecting the current selection before taking on a new selection,
broadcast a deselect packet, so peers deselect it at the same time and
not when receiving another paste packet, fixing the initial issue in a
better way.